Make DataDeps an optional dependency of Comments
Making our code more library friendly is a good idea, since DataDeps
really ties a lot of unrelated things together. However we keep it as a
default fallback in case the `subreddits` param in Comments is
undefined. This avoids us having to refactor a lot more while still
taking the requirement of that depenedncy away.

Added TODO note in recommender.py that we should make that code behave
more like a library and depend less on the application state provided by
DataDeps.