Implement naive recommendations
This system will get a user's own ratings and then compare them with everyone else in the matrix (comparing raw rows, no fancy math here). When we find the most similar row (cosine similarity) we make recommendations based on the most similar user.

This crappy collaborative filtering approach is all I wanted to add for "v1" or whatever of recommendations. Will spend more time in the next few days building SVD approach or something like that.

We apply some filtering to our most similar "neighbor" user: get rid of subreddits the target user (one we queried for) has commented on and ones that the neighbor did not comment on. Recommending subreddits the target user commented on is pointless, since they obviously know about that subreddit. Recommending ones their nearest neighbor did not comment on is also pointless because we have no idea if their neighbor would've liked that subreddit (plus the rating is 0 so it's kinda stupid).

Also it's quite common for there to be no recommendations for a user with this system. Of course we only do gaming recommendations right now so we in general exclude a lot of users, but the nearest-neighbor approach kinda proves to be bad even for other "gamer" redditors we looked at.